Move all android support functions into platform/android.rs
