+++ skia-bindings/build_support.rs
@@ -1,6 +1,5 @@
 //! Build support functions for the Rust-Skia library.
 
-pub mod android;
 pub mod binaries_config;
 #[cfg(feature = "binary-cache")]
 pub mod binary_cache;
